+++ app/src/main/java/com/example/myapplication_1/Adapters/EarningsPageAdapter.java
@@ -1,4 +1,4 @@
-package com.example.myapplication_1;
+package com.example.myapplication_1.Adapters;
 
 import android.content.Context;
 import android.graphics.Bitmap;
@@ -16,7 +16,7 @@ import androidx.recyclerview.widget.DefaultItemAnimator;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import com.example.myapplication_1.Adapters.AdapterCellFeatures;
+import com.example.myapplication_1.R;
 
 import ru.osety.amironlibrary.DrawableUtils;
