+++ app/src/main/java/com/example/myapplication_1/Adapters/EmptyChatWithDriverAdapter.java
@@ -1,0 +1,122 @@
+package com.example.myapplication_1.Adapters;
+
+import android.content.Context;
+import android.graphics.drawable.Drawable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import androidx.annotation.ColorInt;
+import androidx.annotation.NonNull;
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.myapplication_1.R;
+
+import ru.osety.amironlibrary.DrawableUtils;
+
+public class EmptyChatWithDriverAdapter {
+    public static class AdapterGridViewMenu extends RecyclerView.Adapter<AdapterGridViewMenu.ViewHolder> {
+
+        private final ItemsMenu[] itemsMenu;
+        private final LayoutInflater layoutInflater;
+        private final Context context;
+
+        public AdapterGridViewMenu(AdapterGridViewMenu.ItemsMenu[] itemsMenu, Context context) {
+            this.itemsMenu = itemsMenu;
+            this.context = context;
+
+            layoutInflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        }
+
+        @NonNull
+        @Override
+        public ViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
+
+            ViewGroup v = (ViewGroup) layoutInflater.inflate(R.layout.cell_empty_chat_driver, viewGroup, false);
+            return new ViewHolder(v);
+
+        }
+
+        @Override
+        public void onBindViewHolder(@NonNull ViewHolder viewHolder, int i) {
+
+            final ItemsMenu _item = itemsMenu[i];
+
+            Drawable ic_background_xml = context.getResources().getDrawable(R.mipmap.icon_car);
+            float _dens = context.getResources().getDisplayMetrics().density;
+
+            Drawable ic_rout_color = DrawableUtils.setTintDrawable(ic_background_xml, _item.getColorBackground());
+
+            int _size = Math.round(_dens * 40);
+
+            viewHolder.title.setText( _item.getTitle() );
+            viewHolder.cv_item.setOnClickListener( new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+
+                    try {
+
+                        _item.getCallBack().call(_item);
+
+                    } catch ( NullPointerException e) {
+                        e.printStackTrace();
+                    }
+
+                }
+            });
+
+        }
+
+        @Override
+        public int getItemCount() {
+            return itemsMenu.length;
+        }
+
+        public static class ItemsMenu {
+
+            public interface CallBack {
+                void call(ItemsMenu itemsMenu);
+            }
+
+            private @ColorInt
+            int colorBackgroundInt;
+            private CallBack callBack;
+            private String title;
+
+            public ItemsMenu(int colorBackgroundRes, CallBack callBack,String title) {
+                this.colorBackgroundInt = colorBackgroundRes;
+                this.callBack = callBack;
+                this.title = title;
+            }
+
+            public CallBack getCallBack() {
+                return callBack;
+            }
+
+            public String getTitle() {
+                return title;
+            }
+
+            public int getColorBackground() {
+                return colorBackgroundInt;
+            }
+
+        }
+
+        class ViewHolder extends RecyclerView.ViewHolder {
+
+            private final TextView title;
+            private final CardView cv_item;
+
+            public ViewHolder(@NonNull View itemView) {
+                super(itemView);
+                cv_item = itemView.findViewById(R.id.cv_cell__empty_chat_with_driver);
+                title = itemView.findViewById(R.id.cv_cell__empty_chat_with_driver_title);
+
+            }
+        }
+
+    }
+}
