+++ app/src/main/java/com/example/myapplication_1/Adapters/PartnersAdapter.java
@@ -1,6 +1,5 @@
 package com.example.myapplication_1.Adapters;
 
-
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.drawable.Drawable;
@@ -16,23 +15,20 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.myapplication_1.R;
 
-import java.util.List;
-
 import ru.osety.amironlibrary.DrawableUtils;
 
-public class AdapterCreateOrder115 {
+public class PartnersAdapter {
 
     public static class AdapterGridViewMenu extends RecyclerView.Adapter<AdapterGridViewMenu.ViewHolder> {
 
+        private final ItemsMenu[] itemsMenu;
         private final LayoutInflater layoutInflater;
         private final Context context;
 
-        private List<ItemsMenu> itemsMenuList;
-
-
-        public AdapterGridViewMenu(List<AdapterGridViewMenu.ItemsMenu> itemsMenuList, Context context) {
-            this.itemsMenuList = itemsMenuList;
+        public AdapterGridViewMenu(AdapterGridViewMenu.ItemsMenu[] itemsMenu, Context context) {
+            this.itemsMenu = itemsMenu;
             this.context = context;
+
             layoutInflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
         }
 
@@ -40,16 +36,14 @@ public class AdapterCreateOrder115 {
         @Override
         public ViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
 
-
-            View view = LayoutInflater.from(viewGroup.getContext()).inflate(R.layout.cell_create_order_1_1_5, viewGroup, false);
-            ViewHolder itemViewHolder = new ViewHolder(view);
-            return itemViewHolder;
+            ViewGroup v = (ViewGroup) layoutInflater.inflate(R.layout.cell_partners, viewGroup, false);
+            return new ViewHolder(v);
         }
 
         @Override
-        public void onBindViewHolder(@NonNull final ViewHolder viewHolder, final int i) {
+        public void onBindViewHolder(@NonNull ViewHolder viewHolder, int i) {
 
-            final ItemsMenu _item = itemsMenuList.get(i);
+            final ItemsMenu _item = itemsMenu[i];
 
             Drawable ic_background_xml = context.getResources().getDrawable(R.mipmap.icon_car);
             float _dens = context.getResources().getDisplayMetrics().density;
@@ -60,24 +54,15 @@ public class AdapterCreateOrder115 {
             Drawable _def_draw = context.getResources().getDrawable(_item.getImgResId());
             Bitmap _bitmap = DrawableUtils.convertToBitmap(_def_draw, _size, _size);
 
-            viewHolder.title.setText(_item.getStr());
-
-
-            final int currentPosition = i;
-
-            viewHolder.img.setOnClickListener(new View.OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    itemsMenuList.remove(i);
-                    notifyItemRemoved(i);
-                    notifyItemRangeChanged(i, 1);
-                }
-            });
+            viewHolder.img.setBackground( ic_rout_color );
+            viewHolder.img.setScaleType(ImageView.ScaleType.CENTER);
+            viewHolder.img.setImageBitmap( _bitmap );
+            viewHolder.desc.setText( _item.getStr() );
         }
 
         @Override
         public int getItemCount() {
-            return itemsMenuList.size();
+            return itemsMenu.length;
         }
 
         public static class ItemsMenu {
@@ -85,14 +70,13 @@ public class AdapterCreateOrder115 {
             private @ColorInt
             int colorBackgroundInt;
             private int imgResId;
-            private String title;
-            private int imgResId1;
+            private String desc;
 
-            public ItemsMenu(int colorBackgroundRes, int imgResId, String title, int imgResId1) {
+            public ItemsMenu(int colorBackgroundRes,int imgResId,String desc) {
                 this.colorBackgroundInt = colorBackgroundRes;
                 this.imgResId = imgResId;
-                this.title = title;
-                this.imgResId1 = imgResId1;
+                this.desc = desc;
+                this.imgResId = imgResId;
             }
 
             public int getImgResId() {
@@ -100,11 +84,7 @@ public class AdapterCreateOrder115 {
             }
 
             public String getStr() {
-                return title;
-            }
-
-            public int getImgResId1() {
-                return imgResId1;
+                return desc;
             }
 
             public int getColorBackground() {
@@ -115,14 +95,12 @@ public class AdapterCreateOrder115 {
         class ViewHolder extends RecyclerView.ViewHolder {
 
             private final ImageView img;
-            private final TextView title;
-            private final ImageView img1;
+            private final TextView desc;
 
             public ViewHolder(@NonNull View itemView) {
                 super(itemView);
-                img = itemView.findViewById(R.id.rl_cell_create_order_1_1_5_button_cross);
-                title = itemView.findViewById(R.id.rl_cell_create_order_1_1_5_title);
-                img1 = itemView.findViewById(R.id.rl_cell_create_order_1_1_5_button_line);
+                img = itemView.findViewById(R.id.img);
+                desc = itemView.findViewById(R.id.desc);
             }
         }
     }
