+++ app/src/main/java/com/example/myapplication_1/MainActivityCountryCodeSelection.java
@@ -1,26 +1,18 @@
 package com.example.myapplication_1;
 
 import android.content.Context;
-import android.graphics.Bitmap;
-import android.graphics.drawable.Drawable;
 import android.os.Bundle;
 import android.util.AttributeSet;
-import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.ImageButton;
-import android.widget.ImageView;
-import android.widget.TextView;
 
-import androidx.annotation.ColorInt;
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.DefaultItemAnimator;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import ru.osety.amironlibrary.DrawableUtils;
+import com.example.myapplication_1.Adapters.CountryCodeSelectionAdapter;
 
 public class MainActivityCountryCodeSelection extends AppCompatActivity {
 
@@ -63,9 +55,9 @@ public class MainActivityCountryCodeSelection extends AppCompatActivity {
 
         try {
 
-            AdapterGridViewMenu.ItemsMenu[] itemsMenu = getMenuItems();//model_data
+            CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu[] itemsMenu = getMenuItems();//model_data
 
-            AdapterGridViewMenu adapterGridViewMenu = new AdapterGridViewMenu(itemsMenu, getBaseContext());//this;
+            CountryCodeSelectionAdapter.AdapterGridViewMenu adapterGridViewMenu = new CountryCodeSelectionAdapter.AdapterGridViewMenu(itemsMenu, getBaseContext());//this;
             recyclerViewMenu.setAdapter( adapterGridViewMenu );
             recyclerViewMenu.setLayoutManager(
                     new LinearLayoutManager( getBaseContext(), RecyclerView.VERTICAL, false ) );
@@ -77,18 +69,18 @@ public class MainActivityCountryCodeSelection extends AppCompatActivity {
 
     }
 
-    private AdapterGridViewMenu.ItemsMenu[] getMenuItems() {
+    private CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu[] getMenuItems() {
 
-        AdapterGridViewMenu.ItemsMenu []_arr = new AdapterGridViewMenu.ItemsMenu[]{
+        CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu []_arr = new CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu[]{
 
-                new AdapterGridViewMenu.ItemsMenu(
+                new CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu(
                         getResources().getColor(R.color.my_gray),
                         "Г",
                         R.mipmap.georgia,
                         "Грузия",
-                        new AdapterGridViewMenu.ItemsMenu.CallBack() {
+                        new CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu.CallBack() {
                             @Override
-                            public void call(AdapterGridViewMenu.ItemsMenu itemsMenu) {
+                            public void call(CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu itemsMenu) {
 
                                 try {
 
@@ -100,13 +92,13 @@ public class MainActivityCountryCodeSelection extends AppCompatActivity {
 
                             }
                         }, "+995"),
-                new AdapterGridViewMenu.ItemsMenu(
+                new CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu(
                         getResources().getColor(R.color.my_gray),"Р",
                         R.mipmap.russia,
                         "Россия",
-                        new AdapterGridViewMenu.ItemsMenu.CallBack() {
+                        new CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu.CallBack() {
                             @Override
-                            public void call(AdapterGridViewMenu.ItemsMenu itemsMenu) {
+                            public void call(CountryCodeSelectionAdapter.AdapterGridViewMenu.ItemsMenu itemsMenu) {
 
                                 try {
 
@@ -125,119 +117,6 @@ public class MainActivityCountryCodeSelection extends AppCompatActivity {
     }
 
 
-    public static class AdapterGridViewMenu extends RecyclerView.Adapter<AdapterGridViewMenu.ViewHolder> {
 
-        private final ItemsMenu[] itemsMenu;
-        private final LayoutInflater layoutInflater;
-        private final Context context;
-
-        public AdapterGridViewMenu(AdapterGridViewMenu.ItemsMenu[] itemsMenu, Context context) {
-            this.itemsMenu = itemsMenu;
-            this.context = context;
-
-            layoutInflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        }
-
-        @NonNull
-        @Override
-        public ViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
-
-            ViewGroup v = (ViewGroup) layoutInflater.inflate(R.layout.cell_country_code_selection, viewGroup, false);
-            return new ViewHolder(v);
-
-        }
-
-        @Override
-        public void onBindViewHolder(@NonNull ViewHolder viewHolder, int i) {
-
-            final ItemsMenu _item = itemsMenu[i];
-
-            Drawable ic_background_xml = context.getResources().getDrawable(R.mipmap.icon_car);
-            float _dens = context.getResources().getDisplayMetrics().density;
-
-            Drawable ic_rout_color = DrawableUtils.setTintDrawable(ic_background_xml, _item.getColorBackground());
-
-            int _size = Math.round(_dens * 40);
-            Drawable _def_draw = context.getResources().getDrawable(_item.getImgResId());
-            Bitmap _bitmap = DrawableUtils.convertToBitmap(_def_draw, _size, _size);
-
-            viewHolder.img.setBackground( ic_rout_color );
-            viewHolder.img.setScaleType(ImageView.ScaleType.CENTER);
-            viewHolder.img.setImageBitmap( _bitmap );
-            viewHolder.letter.setText( _item.getLetter() );
-            viewHolder.title.setText( _item.getStr() );
-            viewHolder.time.setText( _item.getTime() );
-
-        }
-
-        @Override
-        public int getItemCount() {
-            return itemsMenu.length;
-        }
-
-        public static class ItemsMenu {
-
-            public interface CallBack {
-                void call(ItemsMenu itemsMenu);
-            }
-
-            private @ColorInt int colorBackgroundInt;
-            private String letter;
-            private int imgResId;
-            private String str;
-            private CallBack callBack;
-            private String time;
-
-            public ItemsMenu(int colorBackgroundRes, String letter, int imgResId, String str, CallBack callBack,String time) {
-                this.colorBackgroundInt = colorBackgroundRes;
-                this.letter = letter;
-                this.imgResId = imgResId;
-                this.str = str;
-                this.callBack = callBack;
-                this.time = time;
-            }
-
-            public CallBack getCallBack() {
-                return callBack;
-            }
-
-            public String getLetter(){return letter;}
-
-            public int getImgResId() {
-                return imgResId;
-            }
-
-            public String getStr() {
-                return str;
-            }
-
-            public String getTime() {
-                return time;
-            }
-
-            public int getColorBackground() {
-                return colorBackgroundInt;
-            }
-
-        }
-
-        class ViewHolder extends RecyclerView.ViewHolder {
-
-            private final TextView letter;
-            private final ImageView img;
-            private final TextView title;
-            private final TextView time;
-
-            public ViewHolder(@NonNull View itemView) {
-                super(itemView);
-                letter = itemView.findViewById(R.id.ll_cell_country_code_selection_letter);
-                img = itemView.findViewById(R.id.ll_cell_country_code_selection_img_georgia);
-                title = itemView.findViewById(R.id.ll_cell_country_code_selection_country_georgia);
-                time = itemView.findViewById(R.id.ll_cell_country_code_selection_georgia_number_code);
-
-            }
-        }
-
-    }
 }
