+++ app/src/main/res/layout/car_card.xml
@@ -45,7 +45,7 @@
                             android:scaleType="center"
                             android:scaleX="5"
                             android:scaleY="5"
-                            android:src="@mipmap/red_car" />
+                            android:src="@mipmap/yellow_car" />
 
                         <androidx.coordinatorlayout.widget.CoordinatorLayout
                             android:layout_width="match_parent"
@@ -72,7 +72,7 @@
                                     android:textColor="#000000" />
 
                                     <TextView
-                                        android:id="@+id/time"
+                                        android:id="@+id/cost"
                                         android:layout_width="90dp"
                                         android:layout_height="match_parent"
                                         android:paddingStart="8dp"
@@ -80,7 +80,7 @@
                                         android:paddingEnd="8dp"
                                         android:textSize="14sp"
                                         android:paddingRight="8dp"
-                                        android:text="5 мин"
+                                        android:text="48 ₽"
                                         android:textColor="#000000" />
 
                             </LinearLayout>
