+++ app/src/main/res/layout/cell_settings.xml
@@ -7,7 +7,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:paddingRight="10dp"
-        android:paddingLeft="15dp"
+        android:paddingLeft="20dp"
         android:orientation="horizontal">
         <TextView
             android:id="@+id/desc"
