+++ app/src/main/res/layout/cell_travel_story.xml
@@ -109,14 +109,15 @@
                         android:orientation="horizontal">
 
                         <LinearLayout
-                            android:layout_width="wrap_content"
+                            android:layout_width="match_parent"
                             android:layout_height="wrap_content"
                             android:paddingRight="10dp"
+                            android:layout_weight="1"
                             android:paddingLeft="15dp">
 
                             <Button
                                 android:id="@+id/button_write"
-                                android:layout_width="180dp"
+                                android:layout_width="170dp"
                                 android:layout_height="70dp"
                                 android:drawableStart="@drawable/icon_yellow_help"
                                 android:text="Помощь"
@@ -131,14 +132,15 @@
                         </LinearLayout>
 
                         <LinearLayout
-                            android:layout_width="wrap_content"
+                            android:layout_width="match_parent"
                             android:layout_height="wrap_content"
                             android:paddingRight="10dp"
+                            android:layout_weight="1"
                             android:paddingLeft="5dp">
 
                             <Button
                                 android:id="@+id/button_call"
-                                android:layout_width="180dp"
+                                android:layout_width="170dp"
                                 android:layout_height="70dp"
                                 android:drawableStart="@drawable/icon_phone"
                                 android:text="Водитель"
